Shorten some names.
